requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-mail==1.5.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.5
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
